pyproject.toml
[project]
name = "stock-analysis-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "black>=23.0.0",
    "colorlog>=6.8.0",
    "fastapi>=0.109.0",
    "flake8>=6.1.0",
    "google-adk[a2a]>=0.1.0",
    "google-generativeai>=0.8.0",
    "httpx>=0.27.0",
    "ipykernel>=6.25.0",
    "jupyter>=1.0.0",
    "lxml>=4.9.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "polygon-api-client>=1.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "scikit-learn>=1.3.0",
    "seaborn>=0.13.0",
    "starlette>=0.36.0",
    "ta-lib>=0.4.28",
    "uvicorn>=0.27.0",
    "xgboost>=2.0.0",
    "yfinance>=0.2.40",
]
